requirements.txt
fastapi
uvicorn[standard]
pydantic
email-validator
psycopg2-binary
python-dotenv